[ micro ] lookup `xs` in `open ys` by looking up `ys ++ xs`
Opening module `ys` just means that when we lookup `xs` in the scope
and come across `open ys`, we proceed to lookup `ys ++ xs` (as one
alternative).

Now Name and LName are also indexed by a Scope, and when checking
whether a Name is declared in a C.Decl, we need to check the scope
above in case of `open`.

Unfortunately, the termination checker now complains, as the
termination measure is no longer structural, but it is the combined
size of the scope above and the declarations we are currently
traversing.

Also the continuation-based lookup hasn't been updated (it's
complicated with the scope weakenings), so the test case Shadowed.agda
is broken again.